requirements.txt
asarPy==1.0.1
pillow==11.2.1
pycairo==1.28.0
PyGObject==3.52.3
pystray==0.19.5
python-xlib==0.33
six==1.17.0
